transition: review cleanup — d3-matching start/tween order, per-operator files, comment deslop
