+++ sample/shared/src/commonTest/kotlin/AnimatedBarChartTest.kt
@@ -25,8 +25,6 @@ class AnimatedBarChartTest {
     @Test
     fun barsAnimateFromBaselineToTargets() {
         val root = RootElement()
-
-        // First frame: build the chart and start transitions from the baseline.
         root.tickTransitions(0)
         animatedBarChart(root, width, height, data)
         root.tickTransitions(0)
@@ -34,10 +32,8 @@ class AnimatedBarChartTest {
         val bars = root.bars()
         assertTrue(bars.size == data.size, "expected ${data.size} bars, got ${bars.size}")
 
-        val innerHeight = height - 24f // matches the chart's margin of 12 on each side.
-        val baselineTops = bars.map { it.top }
-        // At the very start every bar sits at the baseline (zero height).
-        baselineTops.forEach { top -> assertTrue(top == innerHeight, "bar should start at baseline") }
+        val innerHeight = height - 24f // The chart applies a margin of 12 on each side.
+        bars.forEach { bar -> assertTrue(bar.top == innerHeight, "bar should start at baseline") }
         assertTrue(root.hasPendingTransitions)
 
         // Midway through the first bar's animation (duration 750ms, no delay for index 0).
@@ -46,11 +42,10 @@ class AnimatedBarChartTest {
         assertTrue(midTop < innerHeight, "bar should have grown above the baseline by the midpoint")
         assertTrue(midTop > 0f)
 
-        // Well past the end of all staggered animations (last delay 90ms + 750ms duration).
+        // Past the end of all staggered animations (last delay 90ms + 750ms duration).
         root.tickTransitions(2000)
         assertFalse(root.hasPendingTransitions, "animations should have completed")
 
-        // Final heights should match the y-scale targets: taller bars for larger values (smaller top).
         val finalTops = root.bars().map { it.top }
         for (i in 1 until finalTops.size) {
             assertTrue(finalTops[i] < finalTops[i - 1], "larger values should produce taller bars")
@@ -63,12 +58,10 @@ class AnimatedBarChartTest {
         root.tickTransitions(0)
         animatedBarChart(root, width, height, data)
         root.tickTransitions(0)
-        root.tickTransitions(2000) // settle
+        root.tickTransitions(2000)
 
         val settledTops = root.bars().map { it.top }
 
-        // A new data set should re-animate the existing bars (no new elements created).
-        root.tickTransitions(2000)
         animatedBarChart(root, width, height, listOf(1.0f, 0.75f, 0.5f, 0.25f))
         root.tickTransitions(2000)
         assertTrue(root.bars().size == data.size, "bar count should be stable across updates")
