+++ transition/api/android/transition.api
@@ -9,6 +9,14 @@ public final class com/juul/krayon/transition/AttributeKt {
 	public static final fun attributeIntOf (Lcom/juul/krayon/transition/Transition;Lkotlin/reflect/KMutableProperty1;Lkotlin/jvm/functions/Function2;)Lcom/juul/krayon/transition/Transition;
 }
 
+public final class com/juul/krayon/transition/CallKt {
+	public static final fun call (Lcom/juul/krayon/transition/Transition;Lkotlin/jvm/functions/Function1;)Lcom/juul/krayon/transition/Transition;
+}
+
+public final class com/juul/krayon/transition/EachKt {
+	public static final fun each (Lcom/juul/krayon/transition/Transition;Lkotlin/jvm/functions/Function2;)Lcom/juul/krayon/transition/Transition;
+}
+
 public final class com/juul/krayon/transition/EaseKt {
 	public static final fun easeBackIn (F)Lcom/juul/krayon/transition/Easing;
 	public static synthetic fun easeBackIn$default (FILjava/lang/Object;)Lcom/juul/krayon/transition/Easing;
@@ -54,21 +62,22 @@ public abstract interface class com/juul/krayon/transition/Easing {
 	public abstract fun ease (F)F
 }
 
+public final class com/juul/krayon/transition/FilterKt {
+	public static final fun filter (Lcom/juul/krayon/transition/Transition;Lkotlin/jvm/functions/Function2;)Lcom/juul/krayon/transition/Transition;
+}
+
 public final class com/juul/krayon/transition/InterruptKt {
 	public static final fun interrupt (Lcom/juul/krayon/selection/Selection;Ljava/lang/String;)Lcom/juul/krayon/selection/Selection;
 	public static final fun interrupt (Lcom/juul/krayon/transition/Transition;)Lcom/juul/krayon/transition/Transition;
 	public static synthetic fun interrupt$default (Lcom/juul/krayon/selection/Selection;Ljava/lang/String;ILjava/lang/Object;)Lcom/juul/krayon/selection/Selection;
 }
 
-public final class com/juul/krayon/transition/OnKt {
-	public static final fun on (Lcom/juul/krayon/transition/Transition;Lcom/juul/krayon/transition/TransitionEvent;Lkotlin/jvm/functions/Function1;)Lcom/juul/krayon/transition/Transition;
+public final class com/juul/krayon/transition/MergeKt {
+	public static final fun merge (Lcom/juul/krayon/transition/Transition;Lcom/juul/krayon/transition/Transition;)Lcom/juul/krayon/transition/Transition;
 }
 
-public final class com/juul/krayon/transition/OperatorsKt {
-	public static final fun call (Lcom/juul/krayon/transition/Transition;Lkotlin/jvm/functions/Function1;)Lcom/juul/krayon/transition/Transition;
-	public static final fun each (Lcom/juul/krayon/transition/Transition;Lkotlin/jvm/functions/Function2;)Lcom/juul/krayon/transition/Transition;
-	public static final fun filter (Lcom/juul/krayon/transition/Transition;Lkotlin/jvm/functions/Function2;)Lcom/juul/krayon/transition/Transition;
-	public static final fun merge (Lcom/juul/krayon/transition/Transition;Lcom/juul/krayon/transition/Transition;)Lcom/juul/krayon/transition/Transition;
+public final class com/juul/krayon/transition/OnKt {
+	public static final fun on (Lcom/juul/krayon/transition/Transition;Lcom/juul/krayon/transition/TransitionEvent;Lkotlin/jvm/functions/Function1;)Lcom/juul/krayon/transition/Transition;
 }
 
 public final class com/juul/krayon/transition/RemoveKt {
