+++ transition/src/commonMain/kotlin/Attribute.kt
@@ -11,12 +11,12 @@ import kotlin.jvm.JvmName
 import kotlin.reflect.KMutableProperty1
 
 /**
- * Animates the [property] of each selected element towards a constant [target] value. The starting
- * value is read from each element when the transition starts, so re-created transitions interrupt and
- * resume smoothly from the current value.
+ * Animates the [property] of each selected element towards a [target] value. The starting value is
+ * read from each element when the transition starts, so re-created transitions interrupt and resume
+ * smoothly from the current value.
  *
- * This is the type-safe analogue of d3's [transition.attr](https://d3js.org/d3-transition/modifying#transition_attr):
- * where d3 uses a string attribute name, Krayon uses a strongly-typed property reference.
+ * This is the type-safe analogue of [d3's transition.attr](https://d3js.org/d3-transition/modifying#transition_attr):
+ * where d3 uses a string attribute name, Krayon uses a property reference.
  */
 @JvmName("attributeFloat")
 public fun <E : Element, D> Transition<E, D>.attribute(
@@ -24,49 +24,49 @@ public fun <E : Element, D> Transition<E, D>.attribute(
     target: Float,
 ): Transition<E, D> = attribute(property) { target }
 
-/** Animates the [property] towards a target computed per element from its datum and index. */
+/** See analogous [d3 function](https://d3js.org/d3-transition/modifying#transition_attr). */
 @JvmName("attributeFloatOf")
 public fun <E : Element, D> Transition<E, D>.attribute(
     property: KMutableProperty1<E, Float>,
     target: E.(Arguments<D, E?>) -> Float,
 ): Transition<E, D> = animateAttribute(property, target) { start, end -> interpolator(start, end) }
 
-/** Animates a [Double] [property] towards a constant [target]. */
+/** See analogous [d3 function](https://d3js.org/d3-transition/modifying#transition_attr). */
 @JvmName("attributeDouble")
 public fun <E : Element, D> Transition<E, D>.attribute(
     property: KMutableProperty1<E, Double>,
     target: Double,
 ): Transition<E, D> = attribute(property) { target }
 
-/** Animates a [Double] [property] towards a target computed per element. */
+/** See analogous [d3 function](https://d3js.org/d3-transition/modifying#transition_attr). */
 @JvmName("attributeDoubleOf")
 public fun <E : Element, D> Transition<E, D>.attribute(
     property: KMutableProperty1<E, Double>,
     target: E.(Arguments<D, E?>) -> Double,
 ): Transition<E, D> = animateAttribute(property, target) { start, end -> interpolator(start, end) }
 
-/** Animates an [Int] [property] towards a constant [target]. */
+/** See analogous [d3 function](https://d3js.org/d3-transition/modifying#transition_attr). */
 @JvmName("attributeInt")
 public fun <E : Element, D> Transition<E, D>.attribute(
     property: KMutableProperty1<E, Int>,
     target: Int,
 ): Transition<E, D> = attribute(property) { target }
 
-/** Animates an [Int] [property] towards a target computed per element. */
+/** See analogous [d3 function](https://d3js.org/d3-transition/modifying#transition_attr). */
 @JvmName("attributeIntOf")
 public fun <E : Element, D> Transition<E, D>.attribute(
     property: KMutableProperty1<E, Int>,
     target: E.(Arguments<D, E?>) -> Int,
 ): Transition<E, D> = animateAttribute(property, target) { start, end -> interpolator(start, end) }
 
-/** Animates a [Color] [property] towards a constant [target], interpolating in ARGB space. */
+/** See analogous [d3 function](https://d3js.org/d3-transition/modifying#transition_attr). */
 @JvmName("attributeColor")
 public fun <E : Element, D> Transition<E, D>.attribute(
     property: KMutableProperty1<E, Color>,
     target: Color,
 ): Transition<E, D> = attribute(property) { target }
 
-/** Animates a [Color] [property] towards a target computed per element. */
+/** See analogous [d3 function](https://d3js.org/d3-transition/modifying#transition_attr). */
 @JvmName("attributeColorOf")
 public fun <E : Element, D> Transition<E, D>.attribute(
     property: KMutableProperty1<E, Color>,
