+++ transition/src/commonMain/kotlin/Remove.kt
@@ -3,8 +3,7 @@ package com.juul.krayon.transition
 import com.juul.krayon.element.Element
 
 /**
- * Removes each selected element from its parent when its transition ends. This is the convenient way
- * to animate an element out (for example, fading or shrinking) before detaching it.
+ * Removes each selected element from its parent when its transition ends.
  *
  * See the analogous [d3 function](https://d3js.org/d3-transition/modifying#transition_remove).
  */
