+++ transition/src/commonMain/kotlin/Timing.kt
@@ -5,17 +5,13 @@ import com.juul.krayon.selection.Arguments
 import com.juul.krayon.selection.Selection
 import com.juul.krayon.selection.each
 
-/**
- * Sets the per-element duration, in milliseconds, for all selected elements.
- *
- * See the analogous [d3 function](https://d3js.org/d3-transition/timing#transition_duration).
- */
+/** See analogous [d3 function](https://d3js.org/d3-transition/timing#transition_duration). */
 public fun <E : Element, D> Transition<E, D>.duration(milliseconds: Long): Transition<E, D> {
     forEachSchedule { it.duration = milliseconds }
     return this
 }
 
-/** Sets the duration per element using [value], evaluated for each element with its datum and index. */
+/** See analogous [d3 function](https://d3js.org/d3-transition/timing#transition_duration). */
 public fun <E : Element, D> Transition<E, D>.duration(
     value: E.(Arguments<D, E?>) -> Long,
 ): Transition<E, D> {
@@ -23,17 +19,13 @@ public fun <E : Element, D> Transition<E, D>.duration(
     return this
 }
 
-/**
- * Sets the per-element start delay, in milliseconds, for all selected elements.
- *
- * See the analogous [d3 function](https://d3js.org/d3-transition/timing#transition_delay).
- */
+/** See analogous [d3 function](https://d3js.org/d3-transition/timing#transition_delay). */
 public fun <E : Element, D> Transition<E, D>.delay(milliseconds: Long): Transition<E, D> {
     forEachSchedule { it.delay = milliseconds }
     return this
 }
 
-/** Sets the delay per element using [value], evaluated for each element with its datum and index. */
+/** See analogous [d3 function](https://d3js.org/d3-transition/timing#transition_delay). */
 public fun <E : Element, D> Transition<E, D>.delay(
     value: E.(Arguments<D, E?>) -> Long,
 ): Transition<E, D> {
@@ -41,11 +33,7 @@ public fun <E : Element, D> Transition<E, D>.delay(
     return this
 }
 
-/**
- * Sets the [Easing] function for all selected elements. Defaults to [easeCubicInOut].
- *
- * See the analogous [d3 function](https://d3js.org/d3-transition/timing#transition_ease).
- */
+/** See analogous [d3 function](https://d3js.org/d3-transition/timing#transition_ease). */
 public fun <E : Element, D> Transition<E, D>.ease(easing: Easing): Transition<E, D> {
     forEachSchedule { it.ease = easing }
     return this
